added subscription validation and created two user API for user

+++ tudime_sms/delete_user_profile.php
@@ -7,8 +7,8 @@ header('Content-Type: application/json; Charset=UTF-8');
 
 
 
-$con=mysqli_connect("localhost","root","IpTv@2019");
-$db=mysqli_select_db($con,"tudime_sms");
+include 'db_config/db_config.php';
+include 'subscription_validation/subscription_validation.php';
 
 $response = array();
 $task = $_POST['task'];
@@ -22,28 +22,30 @@ $id = $_POST['userid'];
 	}
 	else
 	{
-		
-                
-		$sql2 = "Delete FROM user_tbl WHERE `id` ='".$id."'  ";
-		$result_user_histroy = mysqli_query($con,$sql2);
-		mysqli_set_charset($con,"utf8");
-		
-		$sql3 = "Delete FROM tbl_user_profile_image WHERE `user_id` ='".$id."'  ";
-		$result_user_histroy1 = mysqli_query($con,$sql3);
-		mysqli_set_charset($con,"utf8");
-		
-		$sql4 = "Delete FROM chat_backup_tbl WHERE `UserID` ='".$id."'  ";
-		$result_user_histroy2 = mysqli_query($con,$sql4);
-		mysqli_set_charset($con,"utf8");
-		
-		//$sql5 = "Delete FROM buy_tudime_subscription WHERE `useid` ='".$id."'  ";
-		//$result_user_histroy3 = mysqli_query($con,$sql5);
-		//mysqli_set_charset($con,"utf8");
-		
-		
-		
-		$response = array("status" => "success", "error_message" => "", "success_message" => "user profile deleted.", "data" => '');
-		
+		$isSubscriptionValidate = isUserSubscriptionValid($id);
+		if(!$isSubscriptionValidate){
+			$response = array("status" => "error", "error_message" => "useid is subscription", 'success_message' => 'Your subscription has expired, please activate it by purchasing one year subscription.', "data" => "");
+		} else {
+			$sql2 = "Delete FROM user_tbl WHERE `id` ='".$id."'  ";
+			$result_user_histroy = mysqli_query($con,$sql2);
+			mysqli_set_charset($con,"utf8");
+			
+			$sql3 = "Delete FROM tbl_user_profile_image WHERE `user_id` ='".$id."'  ";
+			$result_user_histroy1 = mysqli_query($con,$sql3);
+			mysqli_set_charset($con,"utf8");
+			
+			$sql4 = "Delete FROM chat_backup_tbl WHERE `UserID` ='".$id."'  ";
+			$result_user_histroy2 = mysqli_query($con,$sql4);
+			mysqli_set_charset($con,"utf8");
+			
+			//$sql5 = "Delete FROM buy_tudime_subscription WHERE `useid` ='".$id."'  ";
+			//$result_user_histroy3 = mysqli_query($con,$sql5);
+			//mysqli_set_charset($con,"utf8");
+			
+			
+			
+			$response = array("status" => "success", "error_message" => "", "success_message" => "user profile deleted.", "data" => '');
+		}		
 	}
